pyproject.toml
[project]
name = "krx-data-pipeline"
version = "0.2.0"
description = "Maintainable Korean stock data pipeline: universe sync, daily OHLCV, OpenDART financials/XBRL, KRX flows, and sector-specific operating KPIs."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "KRX Pipeline Team" }]
dependencies = [
    "finance-datareader>=0.9",
    "pykrx>=1.0",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "psycopg2-binary>=2.9",
    "setuptools<70.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "black>=24.0",
    "pytest>=8.0",
]

[project.scripts]
krx-collector = "krx_collector.cli.app:main"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.black]
target-version = ["py312"]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/krx_collector"]
